Add LinkedList.java
This file was provided for us by Professor Vargas. It implements a singly-linked list with internal "Node" and "Iterator" classes. The commentary provided within the file was sparse, but I have chosen to leave it as is. The only addition made to the class was the "size( )" method. Note that this method was implemented with its own internal counter, not with the use of a new instance field external to the method. I made this decision to avoid the necessity of editing the code in several places.